Revert "Removing or supressing build warnings"
This reverts commit ec749dec5c30c65753f1a42938e7c71fb80c3017.

+++ Plugin.CloudFirestore/iOS/ExceptionMapper.cs
@@ -8,7 +8,7 @@ namespace Plugin.CloudFirestore
         public static Exception Map(NSErrorException exception)
         {
             var errorType = ErrorType.Unknown;
-            var errorCode = (FirestoreErrorCode)exception.Error.Code;
+            var errorCode = (FirestoreErrorCode)(long)exception.Error.Code;
 
             switch (errorCode)
             {
